… tests/system_tests/test_functional/metaflow/wine.csv