Speeds up `test_metaflow.py` by:

* (biggest improvement) Adding `wandb.setup()`: each Metaflow step runs in a subprocess, and it takes some time to start `wandb-core`, connect to it, and then shut it down each time
* (surprisingly tiny improvement, maybe even negligible) Simplifying the models and reducing the amount of training data

The tests are still quite slow, with the biggest slowdown being that `wandb.init()` requires logging in which adds ~3s to each step. Another factor is that `@wandb_log` uploads each step's outputs as artifacts, which ends up being slow even with pretty small artifacts (0.5s-1.0s).

Timing changes (comparing this PR to main):

| test | before | after |
| ---- | ---- | ---- |
| test_flow_pytorch | 177s | 132s |
| test_flow_decostep | 119s | 95s |
| test_flow_foreach | 119s | 99s |
| test_flow_decoclass | 110s | 85s |
| test_flow_decoboth | 97s | 83s |

Total improvement is ~128s but it's split between 4 parallel workers (=> 30s). But the tests should use less CPU, which might allow other tests to run faster.